Bitcoin/Dash Arbitrage Trading Bot
This is a simple arbitrage Bot developed in my spare time. The python
file needs to be edited with your specific secret/public keys for
Cryptsy, Bittrex and Poloniex.

The eqDRK and eqBTC fields can easily be adjusted for how much you want
to trade and when you want to refill.